Implemented 3 diffrent ways of puttign images in the app

+++ DiplomaAppV1/app/src/main/java/mom/todorov/elsys/org/diplomaappv1/Place.java
@@ -5,7 +5,7 @@ public class Place {
     String placeName;
     String description;
     String imageLink;
-    int level;
+    long level;
     long typeId;
     String coordinates;
